Simple Bank api

+++ bankAPI/src/main/java/ee/taltech/bankAPI/entity/Bank.java
@@ -18,5 +18,5 @@ public class Bank {
     @NonNull
     private String name;
     @NonNull
-    private Float commission;
+    private Double commission;
 }
